backend/requirements.txt
# Skywalker Backend Dependencies
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
pandas>=2.0.0
numpy>=1.24.0
anthropic>=0.84.0
python-dotenv>=1.0.0
